Multipacket Update
I've updated multipacket array transmission mostly on the Arduino side in order to comply with Kyle's requirements for more than 45 trials. This has broken single packet transmissions because of how I handle packet tracking. PowerBroker2 will probably tell me how to do this better. For now, an issue will be created describing the problem. Planned fix is to introduce a separate set of transmission code for single packets which is pretty hacky...